refactor icons to new naming scheme
