+++ frontend/src/components/shared/buttons/add-btn.component.tsx
@@ -1,4 +1,4 @@
-import AddItemIcon from "../icons/AddItemIcon";
+import { AddItemIcon } from "../icons";
 
 function AddBtn({
   clickSideEffect,
