+++ frontend/src/components/shared/buttons/delete-btn.component.tsx
@@ -1,4 +1,4 @@
-import TrashIcon from "../icons/TrashIcon";
+import { TrashIcon } from "../icons";
 
 function DeleteBtn({
   clickSideEffect,
