+++ frontend/src/components/shared/buttons/delete-person-btn.component.tsx
@@ -1,4 +1,4 @@
-import DeletePersonIcon from "../icons/DeletePersonIcon";
+import { DeletePersonIcon } from "../icons";
 
 function DeletePersonBtn({
   clickSideEffect,
