+++ frontend/src/components/shared/icons/add-item-icon.component.tsx
@@ -14,4 +14,4 @@ function AddItemIcon({ className }: { className: string }) {
   );
 }
 
-export default AddItemIcon;
+export { AddItemIcon };
