+++ frontend/src/components/shared/icons/check-icon.component.tsx
@@ -14,4 +14,4 @@ function CheckIcon({ className }: { className: string }) {
   );
 }
 
-export default CheckIcon;
+export { CheckIcon };
