+++ frontend/src/components/shared/icons/delete-person-icon.component.tsx
@@ -14,4 +14,4 @@ function DeletePersonIcon({ className }: { className: string }) {
   );
 }
 
-export default DeletePersonIcon;
+export { DeletePersonIcon };
