+++ frontend/src/components/shared/icons/external-link-icon.component.tsx
@@ -14,4 +14,4 @@ function ExternalLinkIcon({ className }: { className: string }) {
   );
 }
 
-export default ExternalLinkIcon;
+export { ExternalLinkIcon };
