+++ frontend/src/components/shared/icons/index.tsx
@@ -1,0 +1,12 @@
+export { AddItemIcon } from "./add-item-icon.component";
+export { AddPersonIcon } from "./add-person-icon.component";
+export { CheckIcon } from "./check-icon.component";
+export { CopyIcon } from "./copy-icon.component";
+export { DeletePersonIcon } from "./delete-person-icon.component";
+export { ExternalLinkIcon } from "./external-link-icon.component";
+export { MailIcon } from "./mail-icon.component";
+export { PencilIcon } from "./pencil-icon.component";
+export { ResetIcon } from "./reset-icon.component";
+export { ShareIcon } from "./share-icon.component";
+export { TrashIcon } from "./trash-icon.component";
+export { XIcon } from "./x-icon.component";
