+++ frontend/src/components/shared/icons/mail-icon.component.tsx
@@ -14,4 +14,4 @@ function MailIcon({ className }: { className: string }) {
   );
 }
 
-export default MailIcon;
+export { MailIcon };
