+++ frontend/src/components/shared/icons/share-icon.component.tsx
@@ -14,4 +14,4 @@ function ShareIcon({ className }: { className: string }) {
   );
 }
 
-export default ShareIcon;
+export { ShareIcon };
